+++ src/components/forms/CheckboxGroup/CheckboxGroup.tsx
@@ -8,8 +8,12 @@ import { listStyles, listItemStyles, innerCheckboxStyles, innerCheckboxStylesDis
 
 export const CheckboxGroup: FC<CheckboxGroupProps> = props => {
   const [values, setValues] = useState(props.value || [])
-
   const { onValueChange = () => {}, isDisabled, onBlur: handleBlur, onFocus: handleFocus } = props
+
+  useEffect(() => {
+    setValues(props.value || [])
+  }, [props.value])
+
   useEffect(() => {
     onValueChange(values)
   }, [values])